testing payment

+++ modules/order/OrderService.js
@@ -3,22 +3,27 @@ import CartModel from "../cart/CartModel.js";
 import OrderModel from "./OrderModel.js";
 import PaymentController from "../payment/PaymentController.js"
 import LibraryModel from "../library/LibraryModel.js";
+import UserModel from "../auth/UserModel.js"
 
 const paymentController = new PaymentController()
 
 class OrderService {
 
-    async createOrder (cartId,bookId) {
-
+    async createOrder ({cartId,user,bookId}) {
+        console.log(cartId,bookId,user);
         if(bookId){
+            const boughtByUser = await LibraryModel.findOne({user,book:bookId});
+            if(boughtByUser) return false;
             let book = await BookModel.findById({_id:bookId});
             if(!book)return false;
-            const {email,fullname} = await UserModel.findById({_id:req.user.userId});
-            const {data} = await paymentController.initializePayment({amount:total,email,name:fullname});
-            await OrderModel.create({item:[book],orderValue:book.price,total:book.price,discount:book.discount,paymentReference: data.reference});
+            const {email,fullname} = await UserModel.findById({_id:user});
+            const {data} = await paymentController.initializePayment({amount:book.price,email,name:fullname});
+            
+            await OrderModel.create({user:user,items:[bookId],orderValue:book.price,total:book.price,discount:book.discount,paymentReference: data.reference});
+            
             return data;
         }
-
+        console.log("ss");
         let cart = await CartModel.findById({_id:cartId}).populate("items").populate("user");
         if(!cart)return false;
         const {total} = cart;
@@ -27,12 +32,14 @@ class OrderService {
         const {_id,...cartData} = cart.toJSON();
         await OrderModel.create({...cartData,paymentReference: data.reference});
         return data;
+
     }
 
     async verifyOrder (reference,user,res) {
         const verifyPayment = await paymentController.verifyPayment(reference);
         const order = await OrderModel.findOne({user,paymentReference:reference});
         if(!order) return false;
+        
         if(order.status != "pending"){
             res.status(400);
             throw new Error("Orders has already been attempted for these items.")
@@ -52,8 +59,14 @@ class OrderService {
 
             await Promise.all([...updateBookPromises, ...createLibraryPromises,save]);
 
-            await CartModel.findOneAndDelete({user});
+            if(order.items.length == 1){
+                const cart = await CartModel.findOneAndUpdate({user},{ $pull: { items: order.items[0] } },{new:true});
+                if(!cart.items.length) await cart.remove();
+            } else{
+                await CartModel.findOneAndDelete({user});
+            }
             return true;
+
         } else{
             order.status = "failed";
             await order.save();
